Corrected the .result files (they were mistakenly generated by a wrong version of proj3.py) and added the feature to generate new .result files by changing a variable in test_all.py

+++ proj3/test_all.py
@@ -3,6 +3,9 @@ import sys
 import glob
 import filecmp
 
+# set this to false if you want to generate new *.result files instead of testing
+generate_new = False
+
 files = list(set(glob.glob("graphs/*")) - set(glob.glob("graphs/*.png")) - set(glob.glob("graphs/*.exclude")))
 prefix_lenght = len("graphs/")
 
@@ -10,10 +13,14 @@ for i, val in enumerate(files):
 	files[i] = files[i][prefix_lenght:]
 
 for filename in files:
-	os.system("python proj3.py graphs/" + filename + " --disable-prompt > results/" + filename + ".temp")
+	if not generate_new:
+		os.system("python proj3.py graphs/" + filename + " --disable-prompt > results/" + filename + ".temp")
 	
-	if not filecmp.cmp("results/" + filename + ".result", "results/" + filename + ".temp"):
-		print filename + " returned a different output"
+		if not filecmp.cmp("results/" + filename + ".result", "results/" + filename + ".temp"):
+			print filename + " returned a different output"
+		else:
+			os.system("rm results/" + filename + ".temp")		
 	
-	os.system("rm results/" + filename + ".temp")
+	else:
+		os.system("python proj3.py graphs/" + filename + " --disable-prompt > results/" + filename + ".result")
